Modularize text generators
We can now have multiple of them! The one that self-reports the highest
rating for a given input string will be chosen.
