+++ classes/text_generator.py
